+++ src/reports/patientReportPdfUpdated.js
@@ -1,0 +1,509 @@
+import { bloodpressureQR, bmiQR, tempQR } from 'src/icons/QRCodes'
+import updatedLogo from 'src/icons/UpdatedIcon'
+import { normalizeLangName, parseFromLangKey, setLangUpdated } from '../api/langutil'
+import pdfMake from './pdfMake'
+
+function calculateBMI(heightInCm, weightInKg) {
+  const height = heightInCm / 100
+  const bmi = (weightInKg / height / height).toFixed(1)
+
+  return bmi
+}
+export function generate_pdf_updated(
+  reg,
+  patients,
+  cancer,
+  phlebotomy,
+  fit,
+  wce,
+  doctorSConsult,
+  socialService,
+  geriMmse,
+  geriVision,
+  geriAudiometry,
+  dietitiansConsult,
+  oralHealth,
+  triage,
+  vaccine,
+  lung,
+  nkf,
+  hsg,
+  grace,
+  hearts,
+  geriPtConsult,
+  geriOtConsult,
+  mental,
+  social,
+  podiatry,
+  mammobus,
+  hpv,
+) {
+  console.log('TRIAGE', triage)
+  const language = normalizeLangName(reg?.registrationQ14)
+  const reportFont =
+    language === 'tamil' ? 'NotoTamil' : language === 'mandarin' ? 'PingFangSC' : 'Roboto'
+
+  setLangUpdated(language)
+  let content = []
+
+  content.push(...patientSection(reg, patients))
+  content.push(...temperatureSection(triage))
+  content.push(...bloodPressureSection(triage))
+  content.push(...bmiSection(triage.triageQ10, triage.triageQ11, triage.triageQ12))
+  content.push(...otherScreeningModularitiesSection(reg, geriVision, podiatry, vaccine))
+  //content.push({ text: '', pageBreak: 'before' })
+  content.push(
+    ...followUpSection(
+      reg,
+      vaccine,
+      hsg,
+      lung,
+      phlebotomy,
+      fit,
+      wce,
+      nkf,
+      grace,
+      hearts,
+      oralHealth,
+      mental,
+      mammobus,
+      hpv,
+      socialService,
+    ),
+  )
+  content.push(
+    ...memoSection(geriAudiometry, dietitiansConsult, geriPtConsult, geriOtConsult, doctorSConsult),
+  )
+  content.push(...recommendationSection())
+
+  let fileName = 'Report.pdf'
+  if (patients.initials) {
+    fileName = patients.initials.split(' ').join('_') + '_Report.pdf'
+  }
+
+  pdfMake.fonts = {
+    // download default Roboto font from cdnjs.com
+    Roboto: {
+      normal:
+        'https://cdnjs.cloudflare.com/ajax/libs/pdfmake/0.1.66/fonts/Roboto/Roboto-Regular.ttf',
+      bold: 'https://cdnjs.cloudflare.com/ajax/libs/pdfmake/0.1.66/fonts/Roboto/Roboto-Medium.ttf',
+      italics:
+        'https://cdnjs.cloudflare.com/ajax/libs/pdfmake/0.1.66/fonts/Roboto/Roboto-Italic.ttf',
+      bolditalics:
+        'https://cdnjs.cloudflare.com/ajax/libs/pdfmake/0.1.66/fonts/Roboto/Roboto-MediumItalic.ttf',
+    },
+
+    NotoTamil: {
+      normal: 'https://cdn.jsdelivr.net/gh/choijiwonsoc/my-fonts@main/NotoSansTamil-Regular.ttf',
+      bold: 'https://cdn.jsdelivr.net/gh/choijiwonsoc/my-fonts@main/NotoSansTamil-Bold.ttf',
+      italics: 'https://cdn.jsdelivr.net/gh/choijiwonsoc/my-fonts@main/NotoSansTamil-Regular.ttf',
+      bolditalics:
+        'https://cdn.jsdelivr.net/gh/choijiwonsoc/my-fonts@main/NotoSansTamil-Regular.ttf',
+    },
+
+    // example of usage fonts in collection
+    PingFangSC: {
+      normal: 'https://cdn.jsdelivr.net/gh/choijiwonsoc/my-fonts@main/NotoSansSC-Regular.ttf',
+      bold: 'https://cdn.jsdelivr.net/gh/choijiwonsoc/my-fonts@main/NotoSansSC-Bold.ttf',
+      italics: 'https://cdn.jsdelivr.net/gh/choijiwonsoc/my-fonts@main/NotoSansSC-Regular.ttf',
+      bolditalics: 'https://cdn.jsdelivr.net/gh/choijiwonsoc/my-fonts@main/NotoSansSC-Regular.ttf',
+    },
+  }
+
+  const docDefinition1 = {
+    content: content,
+    styles: {
+      header: {
+        font: reportFont,
+        fontSize: 16,
+        bold: true,
+        margin: [0, 10, 0, 5],
+      },
+      subheader: {
+        font: reportFont,
+        fontSize: 13,
+        bold: true,
+        margin: [0, 3, 0, 3],
+      },
+      normal: {
+        font: reportFont,
+        fontSize: 10,
+        margin: [0, 0, 0, 4],
+      },
+      italicSmall: {
+        font: reportFont,
+        italics: true,
+        fontSize: 10,
+      },
+    },
+    defaultStyle: {
+      font: reportFont,
+      fontSize: 11,
+    },
+    pageMargins: [40, 60, 40, 60],
+  }
+
+  pdfMake.createPdf(docDefinition1).download(fileName)
+}
+
+function patientSection(reg, patients) {
+  const salutation = reg.registrationQ1 || 'Dear'
+
+  const mainLogo = {
+    image: updatedLogo,
+    width: 150,
+  }
+
+  const title = [{ text: parseFromLangKey('title'), style: 'header' }]
+
+  const thanksNote = [
+    { text: `${parseFromLangKey('dear', salutation, reg.registrationQ2)}`, style: 'normal' },
+    { text: `${parseFromLangKey('intro')}`, style: 'normal' },
+  ]
+
+  return [mainLogo, ...title, ...thanksNote]
+}
+
+export function temperatureSection(triage) {
+  const textSection = [
+    { text: `${parseFromLangKey('temp_title')}`, style: 'subheader' },
+    {
+      text: `${parseFromLangKey('temp_reading')} ${triage.triageQ14} °C.\n`,
+      style: 'normal',
+    },
+    {
+      text: `${parseFromLangKey('temp_tip')}`,
+      style: 'normal',
+    },
+  ]
+
+  const imageSection = [
+    {
+      image: tempQR,
+      width: 60,
+      margin: [0, 0, 0, 5],
+    },
+  ]
+
+  return [
+    {
+      columns: [
+        { width: '*', stack: textSection },
+        { width: 'auto', stack: imageSection, alignment: 'right' },
+      ],
+      columnGap: 13,
+      margin: [0, 10, 0, 10],
+    },
+  ]
+}
+
+export function bloodPressureSection(triage) {
+  const textSection = [
+    { text: parseFromLangKey('bp_title'), style: 'subheader' },
+    {
+      text: `${parseFromLangKey('bp_reading')} ${triage.triageQ7}/${triage.triageQ8} mmHg.\n`,
+      style: 'normal',
+    },
+    { text: `${parseFromLangKey('bp_tip')}`, style: 'normal' },
+  ]
+
+  const imageSection = [
+    {
+      image: bloodpressureQR,
+      width: 60,
+      margin: [0, 0, 0, 5],
+    },
+    {
+      text: 'https://www.healthhub.sg/a-z/diseases-and-conditions/understanding-blood-pressure-readings',
+      style: 'italicSmall',
+      fontSize: 7,
+      color: 'blue',
+      link: 'https://www.healthhub.sg/a-z/diseases-and-conditions/understanding-blood-pressure-readings',
+    },
+  ]
+
+  return [
+    {
+      columns: [
+        { width: '*', stack: textSection },
+        { width: 'auto', stack: imageSection, alignment: 'right' },
+      ],
+      columnGap: 13,
+      margin: [0, 10, 0, 10],
+    },
+  ]
+}
+
+export function bmiSection(height, weight, bmiString) {
+  const bmi = calculateBMI(Number(height), Number(weight))
+
+  const imageSection = [
+    {
+      image: bmiQR,
+      width: 60,
+      margin: [0, 0, 0, 5],
+    },
+    // {
+    //   text: 'https://www.healthhub.sg/live-healthy/weight_putting_me_at_risk_of_health_problems',
+    //   style: 'italicSmall',
+    //   fontSize: 7,
+    //   color: 'blue',
+    //   link: 'https://www.healthhub.sg/live-healthy/weight_putting_me_at_risk_of_health_problems',
+    // },
+  ]
+
+  return [
+    { text: parseFromLangKey('bmi_title'), style: 'subheader' },
+    {
+      text: parseFromLangKey('bmi_reading', height, weight, bmiString),
+      style: 'normal',
+    },
+
+    {
+      columns: [
+        {
+          style: 'tableExample',
+          margin: [0, 5, 0, 5],
+          table: {
+            widths: ['*', '*'],
+            body: [
+              [
+                { text: parseFromLangKey('bmi_tbl_l_header'), style: 'tableHeader', bold: true },
+                { text: parseFromLangKey('bmi_tbl_r_header'), style: 'tableHeader', bold: true },
+              ],
+              ['18.5 - 22.9', parseFromLangKey('bmi_tbl_low')],
+              ['23.0 - 27.4', parseFromLangKey('bmi_tbl_mod')],
+              ['27.5 - 32.4', parseFromLangKey('bmi_tbl_high')],
+              ['32.5 - 37.4', parseFromLangKey('bmi_tbl_vhigh')],
+            ],
+          },
+          layout: {
+            hLineWidth: () => 0.5,
+            vLineWidth: () => 0.5,
+            hLineColor: () => 'black',
+            vLineColor: () => 'black',
+          },
+        },
+        { width: 'auto', stack: imageSection, alignment: 'right' },
+      ],
+    },
+    { text: '', margin: [0, 5] },
+  ]
+}
+
+export function otherScreeningModularitiesSection(reg, eye, podiatry, vaccine) {
+  return [
+    { text: parseFromLangKey('other_title'), style: 'subheader' },
+    { text: `${parseFromLangKey('other_eye')}\n`, style: 'normal' },
+    ...(reg?.registrationQ4 >= 60
+      ? [
+          {
+            columns: [
+              {
+                width: '70%',
+                style: 'tableExample',
+                margin: [0, 5, 0, 5],
+                table: {
+                  widths: ['*', '*', '*'],
+                  body: [
+                    [
+                      { text: '', style: 'tableHeader' },
+                      {
+                        text: parseFromLangKey('other_eye_tbl_l_header'),
+                        style: 'tableHeader',
+                        bold: true,
+                      },
+                      {
+                        text: parseFromLangKey('other_eye_tbl_r_header'),
+                        style: 'tableHeader',
+                        bold: true,
+                      },
+                    ],
+                    [
+                      parseFromLangKey('other_eye_tbl_t_row'),
+                      `6/${eye.OphthalQ3}`,
+                      `6/${eye.OphthalQ4}`,
+                    ],
+                    [
+                      parseFromLangKey('other_eye_tbl_b_row'),
+                      `6/${eye.OphthalQ5}`,
+                      `6/${eye.OphthalQ6}`,
+                    ],
+                  ],
+                },
+                layout: {
+                  hLineWidth: () => 0.5,
+                  vLineWidth: () => 0.5,
+                  hLineColor: () => 'black',
+                  vLineColor: () => 'black',
+                },
+              },
+              {
+                width: '*', // takes remaining space
+                text: '', // or you can add other content here or leave blank
+              },
+            ],
+          },
+          { text: '', margin: [0, 5] },
+          { text: `${parseFromLangKey('other_eye_error')} ${eye.OphthalQ8}\n`, style: 'normal' },
+        ]
+      : []),
+    { text: '', margin: [0, 5] },
+    ...(podiatry?.podiatryQ1 === 'Yes'
+      ? [{ text: `${parseFromLangKey('podiatry_screening_true')}\n`, style: 'normal' }]
+      : []),
+    ...(vaccine?.VAX1 === 'Yes'
+      ? [{ text: `${parseFromLangKey('vaccine_1')}\n`, style: 'normal' }]
+      : []),
+    ...(vaccine?.VAX2 === 'Yes'
+      ? [{ text: `${parseFromLangKey('vaccine_2')}\n`, style: 'normal', margin: [20, 0, 0, 0] }]
+      : []),
+    ...(vaccine?.VAX3 === 'Yes'
+      ? [{ text: `${parseFromLangKey('vaccine_3')}\n`, style: 'normal', margin: [20, 0, 0, 20] }]
+      : []),
+  ]
+}
+
+export function followUpSection(
+  reg,
+  vaccine,
+  hsg,
+  lung,
+  phlebotomy,
+  fit,
+  wce,
+  nkf,
+  grace,
+  geriWhForm,
+  oral,
+  mental,
+  mammobus,
+  hpv,
+  socialService,
+) {
+  let vaccineString = null
+  if (vaccine.VAX1 == 'Yes') {
+    vaccineString = `${parseFromLangKey('fw_vax', vaccine.VAX2)}\n`
+  }
+
+  let hsgString = null
+  if (hsg.HSG1 == 'Yes, I signed up for HSG today') {
+    hsgString = `${parseFromLangKey('fw_hsg')}\n`
+  }
+
+  let lungString = null
+  if (lung.LUNG2 == 'Yes') {
+    lungString = `${parseFromLangKey('fw_lung')}\n`
+  }
+
+  let mammobusString = null
+  if (mammobus.mammobusQ1 == 'Yes') {
+    mammobusString = `${parseFromLangKey('fw_mammobus')}\n`
+  }
+
+  let hpvString = null
+  if (hpv.HPV1 == 'Yes') {
+    hpvString = `${parseFromLangKey('fw_hpv')}\n`
+  }
+
+  let mentalString = null
+  if (mental.SAMH2 == 'Yes') {
+    mentalString = `${parseFromLangKey('fw_samh')}\n`
+  }
+
+  let graceString = null
+  if (grace.GRACE2 == 'Yes') {
+    graceString = `${parseFromLangKey('fw_grace', grace.GRACE3)}\n`
+  }
+
+  let whisperString = null
+  if (geriWhForm.WH1 == 'Yes') {
+    whisperString = `${parseFromLangKey('fw_wh')}\n`
+  }
+
+  let aicString = null
+  if (socialService.socialServiceQ4 == 'Yes') {
+    aicString = `${parseFromLangKey('fw_aic')}\n`
+  }
+
+  let oralString = null
+  if (oral.DENT4 == 'Yes') {
+    oralString = `${parseFromLangKey('fw_dent')}\n`
+  }
+
+  return [
+    { text: parseFromLangKey('fw_title'), style: 'subheader' },
+    { text: parseFromLangKey('fw_intro'), style: 'normal' },
+    //...(vaccineString ? [{ text: vaccineString, style: 'normal' }] : []),
+    ...(hsgString ? [{ text: hsgString, style: 'normal' }] : []),
+    ...(lungString ? [{ text: lungString, style: 'normal' }] : []),
+    ,
+    // ...(phlebotomyString ? [{ text: phlebotomyString, style: 'normal' }] : []),
+    // ...(fitString ? [{ text: fitString, style: 'normal' }] : []),
+    // ...(hpvString ? [{ text: hpvString, style: 'normal' }] : []),
+    // ...(nkfString ? [{ text: nkfString, style: 'normal' }] : []),
+
+    ...(graceString ? [{ text: graceString, style: 'normal' }] : []),
+    ...(oralString ? [{ text: oralString, style: 'normal' }] : []),
+    ...(aicString ? [{ text: aicString, style: 'normal' }] : []),
+    ...(mentalString ? [{ text: mentalString, style: 'normal' }] : []),
+    ...(mammobusString ? [{ text: mammobusString, style: 'normal' }] : []),
+    ...(hpvString ? [{ text: hpvString, style: 'normal' }] : []),
+    //...(whisperString ? [{ text: whisperString, style: 'normal' }] : []),
+    { text: '', margin: [0, 5] },
+    //{ text: parseFromLangKey('fw_empty'), style: 'normal' },
+  ]
+}
+
+export function memoSection(audioData, dietData, ptData, otData, doctorData) {
+  let audio =
+    parseFromLangKey('memo_audio') +
+    parseFromLangKey('memo_audio_1', audioData.AudiometryQ12) +
+    parseFromLangKey('memo_audio_2', audioData.AudiometryQ13)
+
+  let diet = parseFromLangKey('memo_diet') + `${dietData.dietitiansConsultQ4}`
+  if (dietData.dietitiansConsultQ5) {
+    diet += parseFromLangKey(
+      'memo_diet_1',
+      dietData.dietitiansConsultQ5,
+      dietData.dietitiansConsultQ6,
+    )
+  }
+
+  const pt = parseFromLangKey('memo_pt') + `${ptData.geriPtConsultQ1}`
+  const ot = parseFromLangKey('memo_ot') + `${otData.geriOtConsultQ1}`
+  const doctor = parseFromLangKey('memo_doctor') + `${doctorData.doctorSConsultQ3}`
+
+  return [
+    { text: parseFromLangKey('memo_title'), style: 'subheader' },
+    {
+      table: {
+        widths: ['*'],
+        body: [
+          [{ text: diet, style: 'normal' }],
+          [{ text: pt, style: 'normal' }],
+          [{ text: ot, style: 'normal' }],
+          [{ text: audio, style: 'normal' }],
+          [{ text: doctor, style: 'normal' }],
+        ],
+      },
+      layout: {
+        fillColor: () => null,
+        hLineColor: () => '#444',
+        vLineColor: () => '#444',
+      },
+      margin: [0, 0, 0, 10],
+    },
+  ]
+}
+
+export function recommendationSection() {
+  return [
+    { text: parseFromLangKey('rec_title'), style: 'subheader' },
+    { text: `${parseFromLangKey('rec')}\n`, style: 'normal' },
+    { text: '', margin: [0, 5] },
+    { text: parseFromLangKey('disclaimer_title'), style: 'subheader' },
+    { text: `${parseFromLangKey('disclaimer')}\n`, style: 'normal' },
+  ]
+}
+
